refactor: fold VatlyHelpers into the Vatly facade
With Vatly::fake() in place there's no reason for a second static-helper
home. Move findBillable(), findBillableOrFail() and cleanUp() onto the
Vatly facade and delete VatlyHelpers — one brand-named surface for fakes,
host-side lookups and webhook-call cleanup (Cashier-style). fake() now
binds the FakeVatly directly instead of hopping through VatlyHelpers.

No back-compat shim — package is still pre-release.

- src/Facades/Vatly.php — absorbs findBillable/findBillableOrFail/cleanUp
- src/VatlyHelpers.php — removed
- tests + README — use Vatly::fake() / Vatly::findBillable() throughout
